Travel-Agency-BE/app/requirements.txt
asgiref==3.9.1
Django==5.2.6
django-ckeditor==6.7.3
django-cors-headers==4.9.0
django-js-asset==3.1.2
djangorestframework==3.16.1
pillow==11.3.0
sqlparse==0.5.3
tzdata==2025.2
